copy connection name to clipboard if it cannot be linked
Let's assume we have a steam account, the link to the steam profile is
simple: "https://steamcommunity.com/profiles/<uid>".

But in the case of discord we don't have an exact link to the user,
just a username, so it makes more sense to just copy this name to the
clipboard than linkin anywhere on discord.